Removed remaining references to the deleted save_results function. Fixed a couple of lines of documentation that previously referenced save_results. Moved _save_generic_results and _save_total_cost_value functions to the __init__ file of the export module. Trigger their execution through the post_solve function defined there. Changed name of first argument in pre_solve and post_solve functions from 'model' to 'instance', to be consistent with the name of the arguments in each of the individual module's post_solve function. Moved the make_iterable function to the export module as well, since it is only referenced by the _save_generic_results function. Cleaned package imports that are no longer necessary in the utilities module. Reverted proposed addition of command line argument for deterministic output, since it will be posted in a different branch.
